Partially revert the split-objs changes So here's the right thing to do for split-objs: 1 compile with -split-objs to get a dir of .o files 2 ghc-6.6 will automatically link all of these into a single .o   file. For earlier versions of ghc we should do this ourselves 3 link all the combined .o files into the GHCi HSfoo.o library 4 ar all the *split* .o files into the GHC libHSfoo.a library
The difference is crucial. We only get the advantage of
split-objs if we ar all the split ones, not the combined ones.
However it's much easier to link the combined ones for the GHCi
library and split-objs provide no benefit for GHCi .o libraries
anyway.

So this patch reinstates the use of the mk/link-splitobjs.sh as
the $(AR) command when we're using split-objs.