Add finialization from the Gtk+ main loop.
This patch adds a replaces the default destroy function for any GObject that
is created in Gtk+ or any libraries that use Gtk+. These objects are now
finalized using an idle handler that is executed by the Gtk+ main loop.
By being executed by the main loop, the finalizers are run from the same
thread as the main loop and, hence, the objects that hold Xlib or Win32
resources will now free these resources from the thread that normally
calls into Xlib/Win32. This fixes a problem with the -threaded runtime
of ghc in which finalization of objects could happen from other threads
which would cause Xlib errors.