Use the C function to finalize dynamic functions.
This patch changes mkDestoryNotifyPtr to only return the address of the
ghc C function freeHaskellPtr. Thus, any closures passed to Gtk+
functions are now directly freed by a C function rather than calling
back into the Haskell system. The downside is that this C function does not
know which object to finalize. It is, however, called with the data that
is supposed to be passed to the callback function. This used to be
nullPtr since this argument was not used. To make things work,
this argument now always has to be the address of the callback function
so that freeHaskellPtr will get this address as a an argument.
If I forgot to change any uses of mkDestroyNotifyPtr, then using this
function will lead to a SEGFAULT since freeHaskellPtr is called with NULL.
Gtk2Hs programs producing a SEGFAULT when closing are an indication of this.