use forward-compatible Foreign.C.Types imports
Previously, the hierarchy generator used #if in an attempt to choose between
two choices for an import line. The one choice gives warnings on old GHC's; the
other gives warnings on new GHC's. However, the symbol used in the #if isn't
defined by gtk2hs' custom build system, so one of the two choices was the
defacto code. Until this patch, the defacto code was the one that produced
warnings on new GHC's.

It should be possible to revise the build system to define this symbol, but it
is becoming less and less likely I will have the energy to work out how to do
this correctly. Instead this patch makes the choice to prefer the import that
compiles cleanly on new GHC's, and let old GHC's print warnings.